% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_adjust.R
\name{pc_adjust}
\alias{pc_adjust}
\title{Population adjuster}
\usage{
pc_adjust(any_df, state_type)
}
\arguments{
\item{any_df}{The name of the dataframe you want to append to}

\item{state_type}{if state identification is abbreviations, use "Abbrev" if full state name, use "State"}
}
\description{
This function appends state population data
}
\examples{
pc_adjust()
}
\keyword{capita}
\keyword{per}

